CTT Web Server v1.4.0

+++ java/components/validation-ctt/HEAD/ctt_web_server/src/main/java/com/at4wireless/spring/service/TestCaseServiceImpl.java
@@ -120,7 +120,7 @@ public class TestCaseServiceImpl implements TestCaseService {
 
 			if (!condition.isEmpty()) {
 				try {
-					if ((Double)engine.eval(condition)==1.0) {
+					if ((Integer)engine.eval(condition)==1.0) {
 						applyTC.add(tc);
 					}
 				} catch (ScriptException e) {
